+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/service/operation/PermissionCrudOperation.java
@@ -2,9 +2,6 @@ package com.dwarfeng.rbacds.impl.service.operation;
 
 import com.dwarfeng.rbacds.stack.bean.entity.Permission;
 import com.dwarfeng.rbacds.stack.cache.PermissionCache;
-import com.dwarfeng.rbacds.stack.cache.PermissionUserCache;
-import com.dwarfeng.rbacds.stack.cache.RolePermissionCache;
-import com.dwarfeng.rbacds.stack.cache.UserPermissionCache;
 import com.dwarfeng.rbacds.stack.dao.PermissionDao;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes;
 import com.dwarfeng.subgrade.sdk.service.custom.operation.BatchCrudOperation;
@@ -21,25 +18,15 @@ public class PermissionCrudOperation implements BatchCrudOperation<StringIdKey, 
     private final PermissionDao permissionDao;
     private final PermissionCache permissionCache;
 
-    private final UserPermissionCache userPermissionCache;
-    private final RolePermissionCache rolePermissionCache;
-    private final PermissionUserCache permissionUserCache;
-
     @Value("${cache.timeout.entity.permission}")
     private long permissionTimeout;
 
     public PermissionCrudOperation(
             PermissionDao permissionDao,
-            PermissionCache permissionCache,
-            UserPermissionCache userPermissionCache,
-            RolePermissionCache rolePermissionCache,
-            PermissionUserCache permissionUserCache
+            PermissionCache permissionCache
     ) {
         this.permissionDao = permissionDao;
         this.permissionCache = permissionCache;
-        this.userPermissionCache = userPermissionCache;
-        this.rolePermissionCache = rolePermissionCache;
-        this.permissionUserCache = permissionUserCache;
     }
 
     @Override
@@ -63,14 +50,6 @@ public class PermissionCrudOperation implements BatchCrudOperation<StringIdKey, 
 
     @Override
     public StringIdKey insert(Permission permission) throws Exception {
-        // 清空用户权限缓存。
-        userPermissionCache.clear();
-        // 清空角色权限缓存。
-        rolePermissionCache.clear();
-        // 清空权限用户缓存。
-        permissionUserCache.clear();
-
-        // 插入权限自身。
         permissionDao.insert(permission);
         permissionCache.push(permission, permissionTimeout);
         return permission.getKey();
@@ -78,27 +57,12 @@ public class PermissionCrudOperation implements BatchCrudOperation<StringIdKey, 
 
     @Override
     public void update(Permission permission) throws Exception {
-        // 清空用户权限缓存。
-        userPermissionCache.clear();
-        // 清空角色权限缓存。
-        rolePermissionCache.clear();
-        // 清空权限用户缓存。
-        permissionUserCache.clear();
-
-        // 更新权限自身。
         permissionCache.push(permission, permissionTimeout);
         permissionDao.update(permission);
     }
 
     @Override
     public void delete(StringIdKey key) throws Exception {
-        // 清空用户权限缓存。
-        userPermissionCache.clear();
-        // 清空角色权限缓存。
-        rolePermissionCache.clear();
-        // 清空权限用户缓存。
-        permissionUserCache.clear();
-
         // 删除权限自身。
         permissionCache.delete(key);
         permissionDao.delete(key);
@@ -130,43 +94,20 @@ public class PermissionCrudOperation implements BatchCrudOperation<StringIdKey, 
 
     @Override
     public List<StringIdKey> batchInsert(List<Permission> permissions) throws Exception {
-        // 清空用户权限缓存。
-        userPermissionCache.clear();
-        // 清空角色权限缓存。
-        rolePermissionCache.clear();
-        // 清空权限用户缓存。
-        permissionUserCache.clear();
-
-        // 插入权限自身。
         permissionCache.batchPush(permissions, permissionTimeout);
         return permissionDao.batchInsert(permissions);
     }
 
     @Override
     public void batchUpdate(List<Permission> permissions) throws Exception {
-        // 清空用户权限缓存。
-        userPermissionCache.clear();
-        // 清空角色权限缓存。
-        rolePermissionCache.clear();
-        // 清空权限用户缓存。
-        permissionUserCache.clear();
-
-        // 更新权限自身。
         permissionCache.batchPush(permissions, permissionTimeout);
         permissionDao.batchUpdate(permissions);
     }
 
     @Override
     public void batchDelete(List<StringIdKey> keys) throws Exception {
-        // 清空用户权限缓存。
-        userPermissionCache.clear();
-        // 清空角色权限缓存。
-        rolePermissionCache.clear();
-        // 清空权限用户缓存。
-        permissionUserCache.clear();
-
-        // 删除权限自身。
-        permissionCache.batchDelete(keys);
-        permissionDao.batchDelete(keys);
+        for (StringIdKey key : keys) {
+            delete(key);
+        }
     }
 }