+++ rbac-distributed-service-node/src/main/java/com/dwarfeng/rbacds/node/configuration/CacheConfiguration.java
@@ -5,7 +5,6 @@ import com.dwarfeng.rbacds.sdk.bean.entity.*;
 import com.dwarfeng.rbacds.stack.bean.entity.*;
 import com.dwarfeng.subgrade.impl.bean.MapStructBeanTransformer;
 import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
-import com.dwarfeng.subgrade.impl.cache.RedisKeyListCache;
 import com.dwarfeng.subgrade.sdk.redis.formatter.LongIdStringKeyFormatter;
 import com.dwarfeng.subgrade.sdk.redis.formatter.StringIdStringKeyFormatter;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
@@ -28,12 +27,6 @@ public class CacheConfiguration {
     private String rolePrefix;
     @Value("${cache.prefix.entity.user}")
     private String userPrefix;
-    @Value("${cache.prefix.list.user_has_permission}")
-    private String userPermissionListKey;
-    @Value("${cache.prefix.list.role_has_permission}")
-    private String rolePermissionListKey;
-    @Value("${cache.prefix.list.permission_has_user}")
-    private String permissionUserListKey;
     @Value("${cache.prefix.entity.permission_group}")
     private String permissionGroupPrefix;
     @Value("${cache.prefix.entity.filter_support}")
@@ -79,36 +72,6 @@ public class CacheConfiguration {
         return new RedisBatchBaseCache<>(
                 (RedisTemplate<String, FastJsonUser>) template,
                 new StringIdStringKeyFormatter(userPrefix),
-                new MapStructBeanTransformer<>(User.class, FastJsonUser.class, BeanMapper.class)
-        );
-    }
-
-    @Bean
-    @SuppressWarnings("unchecked")
-    public RedisKeyListCache<StringIdKey, Permission, FastJsonPermission> userPermissionRedisKeyListCache() {
-        return new RedisKeyListCache<>(
-                (RedisTemplate<String, FastJsonPermission>) template,
-                new StringIdStringKeyFormatter(userPermissionListKey),
-                new MapStructBeanTransformer<>(Permission.class, FastJsonPermission.class, BeanMapper.class)
-        );
-    }
-
-    @Bean
-    @SuppressWarnings("unchecked")
-    public RedisKeyListCache<StringIdKey, Permission, FastJsonPermission> rolePermissionRedisKeyListCache() {
-        return new RedisKeyListCache<>(
-                (RedisTemplate<String, FastJsonPermission>) template,
-                new StringIdStringKeyFormatter(rolePermissionListKey),
-                new MapStructBeanTransformer<>(Permission.class, FastJsonPermission.class, BeanMapper.class)
-        );
-    }
-
-    @Bean
-    @SuppressWarnings("unchecked")
-    public RedisKeyListCache<StringIdKey, User, FastJsonUser> permissionUserRedisKeyListCache() {
-        return new RedisKeyListCache<>(
-                (RedisTemplate<String, FastJsonUser>) template,
-                new StringIdStringKeyFormatter(permissionUserListKey),
                 new MapStructBeanTransformer<>(User.class, FastJsonUser.class, BeanMapper.class)
         );
     }